Fix test_plugin_request_handling.py
