+++ plenum/test/plugin/demo_plugin/helper.py
@@ -1,6 +1,6 @@
 from plenum.common.constants import TXN_TYPE, DATA
+from plenum.test.helper import sdk_gen_request, sdk_sign_and_submit_req_obj, sdk_get_reply
 from plenum.test.plugin.demo_plugin.constants import AUCTION_START
-from plenum.test.plugin.demo_plugin.test_plugin_request_handling import successful_op
 
 
 def send_auction_txn(looper,
@@ -10,3 +10,10 @@ def send_auction_txn(looper,
         DATA: {'id': 'abc'}
     }
     successful_op(looper, op, sdk_wallet_steward, sdk_pool_handle)
+
+
+def successful_op(looper, op, sdk_wallet, sdk_pool_handle):
+    req_obj = sdk_gen_request(op, identifier=sdk_wallet[1])
+    req = sdk_sign_and_submit_req_obj(looper, sdk_pool_handle,
+                                      sdk_wallet, req_obj)
+    sdk_get_reply(looper, req)
