+++ plenum/test/plugin/demo_plugin/test_catchup.py
@@ -7,8 +7,9 @@ from plenum.test.plugin.demo_plugin import AUCTION_LEDGER_ID
 
 from plenum.test.plugin.demo_plugin.constants import AUCTION_END, PLACE_BID, \
     AMOUNT, AUCTION_START
+from plenum.test.plugin.demo_plugin.helper import successful_op
 from plenum.test.plugin.demo_plugin.test_plugin_request_handling import \
-    some_requests, successful_op
+    some_requests
 from plenum.test.pool_transactions.helper import \
     disconnect_node_and_ensure_disconnected, reconnect_node_and_ensure_connected
